requirements.txt
streamlit==1.39.0
pandas==2.2.2
numpy==1.26.4
python-pptx==0.6.23
openpyxl==3.1.5
